# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/participantes": {
      "get": {
        "operationId": "ParticipanteController_listar",
        "summary": "API para obtener el listado de participantes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      },
      "post": {
        "operationId": "ParticipanteController_crear",
        "summary": "API para crear un nuevo participante",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo participante utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearParticipanteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/{id}": {
      "get": {
        "operationId": "ParticipanteController_obtenerPorId",
        "summary": "API para obtener un participante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      },
      "patch": {
        "operationId": "ParticipanteController_actualizar",
        "summary": "API para actualizar un participante",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un participante existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarParticipanteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/eliminar/{id}": {
      "delete": {
        "operationId": "ParticipanteController_eliminar",
        "summary": "API para eliminar un participante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/soft-delete/{id}": {
      "patch": {
        "operationId": "ParticipanteController_softDelete",
        "summary": "API para eliminar un participante por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/restore/{id}": {
      "patch": {
        "operationId": "ParticipanteController_restore",
        "summary": "API para restaurar un participante eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
...
    "/pistas": {
      "get": {
        "operationId": "PistaController_listar",
        "summary": "API para obtener el listado de pistas",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      },
      "post": {
        "operationId": "PistaController_crear",
        "summary": "API para crear un nuevo pista",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo pista utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPistaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      }
    },
    "/pistas/{id}": {
      "get": {
        "operationId": "PistaController_obtenerPorId",
        "summary": "API para obtener un pista por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del pista que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      },
      "patch": {
        "operationId": "PistaController_actualizar",
        "summary": "API para actualizar un pista",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del pista que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un pista existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPistaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      }
    },
    "/pistas/eliminar/{id}": {
      "delete": {
        "operationId": "PistaController_eliminar",
        "summary": "API para eliminar un pista por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del pista que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      }
    },
    "/pistas/soft-delete/{id}": {
      "patch": {
        "operationId": "PistaController_softDelete",
        "summary": "API para eliminar un pista por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del pista que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      }
    },
    "/pistas/restore/{id}": {
      "patch": {
        "operationId": "PistaController_restore",
        "summary": "API para restaurar un pista eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del pista que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pista"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend",
    "description": "Documentación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "CrearParticipanteDto": {
        "type": "object",
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "example": "Arturo Rodrigo Perez Chavez"
          },
          "carnetIdentidad": {
            "type": "string",
            "example": "8526947"
          },
          "fechaNacimiento": {
            "type": "string",
            "example": "2003-05-23"
          },
          "departamento": {
            "type": "string",
            "example": "La Paz"
          },
          "provincia": {
            "type": "string",
            "example": "Murillo"
          },
          "municipio": {
            "type": "string",
            "example": "Achocalla"
          }
        },
        "required": [
          "nombreCompleto",
          "carnetIdentidad",
          "fechaNacimiento",
          "departamento",
          "provincia",
          "municipio"
        ]
      },
      "ActualizarParticipanteDto": {
        "type": "object",
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "example": "Arturo Rodrigo Perez Chavez"
          },
          "carnetIdentidad": {
            "type": "string",
            "example": "8526947"
          },
          "fechaNacimiento": {
            "type": "string",
            "example": "2003-05-23"
          },
          "departamento": {
            "type": "string",
            "example": "La Paz"
          },
          "provincia": {
            "type": "string",
            "example": "Murillo"
          },
          "municipio": {
            "type": "string",
            "example": "Achocalla"
          }
        },
        "required": [
          "nombreCompleto",
          "carnetIdentidad",
          "fechaNacimiento",
          "departamento",
          "provincia",
          "municipio"
        ]
      },
...
      "CrearPistaDto": {
        "type": "object",
        "properties": {
          "nombrePista": {
            "type": "string",
            "example": "Pista 1"
          }
        },
        "required": [
          "nombrePista"
        ]
      },
      "ActualizarPistaDto": {
        "type": "object",
        "properties": {
          "nombrePista": {
            "type": "string",
            "example": "Pista 1"
          }
        },
        "required": [
          "nombrePista"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}